refactor: moificar estrutura relatorioPDF de fetch para axios
