+++ frontend/atendimento-app/src/features/relatorio/utils/formatarData.ts
@@ -1,0 +1,3 @@
+export function formatarDataBr(dataIso: string): string {
+  return new Date(dataIso).toLocaleDateString("pt-BR");
+}
